+++ plugins/processing/classification/src/algorithms/ovpCAlgorithmLDADiscriminantFunction.h
@@ -25,7 +25,7 @@ namespace OpenViBEPlugins
 
 			//Return the class membership of the feature vector
 			double getValue(const Eigen::VectorXd& featureVector) { return (m_weight.transpose() * featureVector)(0) + m_bias; }
-			uint32_t getWeightVectorSize() const { return uint32_t(m_weight.size()); }
+			uint32_t getWeightVectorSize() const { return m_weight.size(); }
 
 
 			bool loadConfiguration(const XML::IXMLNode* configuration);
